src/requirements.txt
django
django-admin-volt
django-cors-headers
djangorestframework
django-crispy-forms
django_extensions
django-debug-toolbar
django-filter
djLint
markdown